+++ crates/nostr-relay-pool/CHANGELOG.md
@@ -25,6 +25,10 @@
 
 ## Unreleased
 
+### Breaking changes
+
+- Change `RelayPool::stream_events*` output to include the `RelayUrl` and `Result`, enabling callers to identify which relay sent the event or if a specific relay encountered an error (https://github.com/rust-nostr/nostr/pull/1156)
+
 ### Added
 
 - Add `Output::new` constructor and `Output::into_inner` method